+++ cmd/otel-allocator/server/server.go
@@ -16,6 +16,7 @@ package server
 
 import (
 	"context"
+	"crypto/tls"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -32,6 +33,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
+	promcommconfig "github.com/prometheus/common/config"
 	promconfig "github.com/prometheus/prometheus/config"
 	"gopkg.in/yaml.v2"
 
@@ -63,16 +65,46 @@ type Server struct {
 	logger         logr.Logger
 	allocator      allocation.Allocator
 	server         *http.Server
+	httpsServer    *http.Server
 	jsonMarshaller jsoniter.API
 
 	// Use RWMutex to protect scrapeConfigResponse, since it
 	// will be predominantly read and only written when config
 	// is applied.
-	mtx                  sync.RWMutex
-	scrapeConfigResponse []byte
+	mtx                                  sync.RWMutex
+	scrapeConfigResponse                 []byte
+	ScrapeConfigMarshalledSecretResponse []byte
 }
 
-func NewServer(log logr.Logger, allocator allocation.Allocator, listenAddr string) *Server {
+type Option func(*Server)
+
+// Option to create an additional https server with mTLS configuration.
+// Used for getting the scrape config with real secret values.
+func WithTLSConfig(tlsConfig *tls.Config, httpsListenAddr string) Option {
+	return func(s *Server) {
+		httpsRouter := gin.New()
+		s.setRouter(httpsRouter)
+
+		s.httpsServer = &http.Server{Addr: httpsListenAddr, Handler: httpsRouter, ReadHeaderTimeout: 90 * time.Second, TLSConfig: tlsConfig}
+	}
+}
+
+func (s *Server) setRouter(router *gin.Engine) {
+	router.Use(gin.Recovery())
+	router.UseRawPath = true
+	router.UnescapePathValues = false
+	router.Use(s.PrometheusMiddleware)
+
+	router.GET("/scrape_configs", s.ScrapeConfigsHandler)
+	router.GET("/jobs", s.JobHandler)
+	router.GET("/jobs/:job_id/targets", s.TargetsHandler)
+	router.GET("/metrics", gin.WrapH(promhttp.Handler()))
+	router.GET("/livez", s.LivenessProbeHandler)
+	router.GET("/readyz", s.ReadinessProbeHandler)
+	registerPprof(router.Group("/debug/pprof/"))
+}
+
+func NewServer(log logr.Logger, allocator allocation.Allocator, listenAddr string, options ...Option) *Server {
 	s := &Server{
 		logger:         log,
 		allocator:      allocator,
@@ -81,19 +113,14 @@ func NewServer(log logr.Logger, allocator allocation.Allocator, listenAddr strin
 
 	gin.SetMode(gin.ReleaseMode)
 	router := gin.New()
-	router.Use(gin.Recovery())
-	router.UseRawPath = true
-	router.UnescapePathValues = false
-	router.Use(s.PrometheusMiddleware)
-	router.GET("/scrape_configs", s.ScrapeConfigsHandler)
-	router.GET("/jobs", s.JobHandler)
-	router.GET("/jobs/:job_id/targets", s.TargetsHandler)
-	router.GET("/metrics", gin.WrapH(promhttp.Handler()))
-	router.GET("/livez", s.LivenessProbeHandler)
-	router.GET("/readyz", s.ReadinessProbeHandler)
-	registerPprof(router.Group("/debug/pprof/"))
+	s.setRouter(router)
 
 	s.server = &http.Server{Addr: listenAddr, Handler: router, ReadHeaderTimeout: 90 * time.Second}
+
+	for _, opt := range options {
+		opt(s)
+	}
+
 	return s
 }
 
@@ -105,6 +132,16 @@ func (s *Server) Start() error {
 func (s *Server) Shutdown(ctx context.Context) error {
 	s.logger.Info("Shutting down server...")
 	return s.server.Shutdown(ctx)
+}
+
+func (s *Server) StartHTTPS() error {
+	s.logger.Info("Starting HTTPS server...")
+	return s.httpsServer.ListenAndServeTLS("", "")
+}
+
+func (s *Server) ShutdownHTTPS(ctx context.Context) error {
+	s.logger.Info("Shutting down HTTPS server...")
+	return s.httpsServer.Shutdown(ctx)
 }
 
 // RemoveRegexFromRelabelAction is needed specifically for keepequal/dropequal actions because even though the user doesn't specify the
@@ -151,15 +188,14 @@ func RemoveRegexFromRelabelAction(jsonConfig []byte) ([]byte, error) {
 	return jsonConfigNew, nil
 }
 
-// UpdateScrapeConfigResponse updates the scrape config response. The target allocator first marshals these
-// configurations such that the underlying prometheus marshaling is used. After that, the YAML is converted
-// in to a JSON format for consumers to use.
-func (s *Server) UpdateScrapeConfigResponse(configs map[string]*promconfig.ScrapeConfig) error {
-	var configBytes []byte
+func (s *Server) MarshalScrapeConfig(configs map[string]*promconfig.ScrapeConfig, marshalSecretValue bool) error {
+	promcommconfig.MarshalSecretValue = marshalSecretValue
+
 	configBytes, err := yaml.Marshal(configs)
 	if err != nil {
 		return err
 	}
+
 	var jsonConfig []byte
 	jsonConfig, err = yaml2.YAMLToJSON(configBytes)
 	if err != nil {
@@ -172,8 +208,28 @@ func (s *Server) UpdateScrapeConfigResponse(configs map[string]*promconfig.Scrap
 	}
 
 	s.mtx.Lock()
-	s.scrapeConfigResponse = jsonConfigNew
+	if marshalSecretValue {
+		s.ScrapeConfigMarshalledSecretResponse = jsonConfigNew
+	} else {
+		s.scrapeConfigResponse = jsonConfigNew
+	}
 	s.mtx.Unlock()
+
+	return nil
+}
+
+// UpdateScrapeConfigResponse updates the scrape config response. The target allocator first marshals these
+// configurations such that the underlying prometheus marshaling is used. After that, the YAML is converted
+// in to a JSON format for consumers to use.
+func (s *Server) UpdateScrapeConfigResponse(configs map[string]*promconfig.ScrapeConfig) error {
+	err := s.MarshalScrapeConfig(configs, false)
+	if err != nil {
+		return err
+	}
+	err = s.MarshalScrapeConfig(configs, true)
+	if err != nil {
+		return err
+	}
 	return nil
 }
 
@@ -181,6 +237,9 @@ func (s *Server) UpdateScrapeConfigResponse(configs map[string]*promconfig.Scrap
 func (s *Server) ScrapeConfigsHandler(c *gin.Context) {
 	s.mtx.RLock()
 	result := s.scrapeConfigResponse
+	if c.Request.TLS != nil {
+		result = s.ScrapeConfigMarshalledSecretResponse
+	}
 	s.mtx.RUnlock()
 
 	// We don't use the jsonHandler method because we don't want our bytes to be re-encoded