+++ internal/manifests/collector/container.go
@@ -141,8 +141,9 @@ func Container(cfg config.Config, logger logr.Logger, otelcol v1beta1.OpenTeleme
 	}
 
 	var livenessProbe *corev1.Probe
+	var readinessProbe *corev1.Probe
 	if configFromString, err := adapters.ConfigFromString(configYaml); err == nil {
-		if probe, err := getLivenessProbe(configFromString, otelcol.Spec.LivenessProbe); err == nil {
+		if probe, err := getProbe(configFromString, otelcol.Spec.LivenessProbe); err == nil {
 			livenessProbe = probe
 		} else if errors.Is(err, adapters.ErrNoServiceExtensions) {
 			logger.V(4).Info("extensions not configured, skipping liveness probe creation")
@@ -150,6 +151,16 @@ func Container(cfg config.Config, logger logr.Logger, otelcol v1beta1.OpenTeleme
 			logger.V(4).Info("healthcheck extension not configured, skipping liveness probe creation")
 		} else {
 			logger.Error(err, "cannot create liveness probe.")
+		}
+
+		if probe, err := getProbe(configFromString, otelcol.Spec.ReadinessProbe); err == nil {
+			readinessProbe = probe
+		} else if errors.Is(err, adapters.ErrNoServiceExtensions) {
+			logger.V(4).Info("extensions not configured, skipping readiness probe creation")
+		} else if errors.Is(err, adapters.ErrNoServiceExtensionHealthCheck) {
+			logger.V(4).Info("healthcheck extension not configured, skipping readiness probe creation")
+		} else {
+			logger.Error(err, "cannot create readiness probe.")
 		}
 	}
 
@@ -188,6 +199,7 @@ func Container(cfg config.Config, logger logr.Logger, otelcol v1beta1.OpenTeleme
 		Resources:       otelcol.Spec.Resources,
 		SecurityContext: otelcol.Spec.SecurityContext,
 		LivenessProbe:   livenessProbe,
+		ReadinessProbe:  readinessProbe,
 		Lifecycle:       otelcol.Spec.Lifecycle,
 	}
 }
@@ -250,7 +262,7 @@ func portMapToList(portMap map[string]corev1.ContainerPort) []corev1.ContainerPo
 	return ports
 }
 
-func getLivenessProbe(config map[interface{}]interface{}, probeConfig *v1beta1.Probe) (*corev1.Probe, error) {
+func getProbe(config map[interface{}]interface{}, probeConfig *v1beta1.Probe) (*corev1.Probe, error) {
 	probe, err := adapters.ConfigToContainerProbe(config)
 	if err != nil {
 		return nil, err